Add files via upload
code by Geoff Maguda


NB: icomoon.css is nothing much but a css icons. its the one showing the left and right arrows and image thumbnails,
    you may choose to remove its okay.

NB: geoffslider.js is the actual jquery javascript that is processing / or making the slider happen . its the real deal
NB: geofslider.css is just css adding some styles for the page. you can ignore it. unless you understand it. just focus on the html and geoffslider.js to understand the slider.


NB:     i will be going to create video youtube tutorial soon to be uploaded on my channel 'Rectar Learns'. Look out for it. dont worry if you dont understand it.

         check out my youtube channel the week starting 6- sept- 2021

   test  index.html    to see it for yourself